// packages/app/frontend/components/category/types.ts
export type CategoryIconType =
  | 'tag'
  | 'wallet'
  | 'trending_up'
  | 'trending_down'
  | 'piggy_bank'
  | 'house'
  | 'utensils'
  | 'shopping_cart'
  | 'car'
  | 'bus'
  | 'plane'
  | 'heart_pulse'
  | 'graduation_cap'
  | 'briefcase'
  | 'music'
  | 'zap'
  | 'wifi'
  | 'shirt'
  | 'dumbbell'
  | 'coffee'
  | 'gift'
  | 'book'
  | 'baby'
  | 'plus'

export type CategoryColor =
  | 'red'
  | 'orange'
  | 'yellow'
  | 'green'
  | 'teal'
  | 'blue'
  | 'purple'
  | 'pink'

export const CategoryIcons: CategoryIconType[] = [
  'tag',
  'wallet',
  'trending_up',
  'trending_down',
  'piggy_bank',
  'house',
  'utensils',
  'shopping_cart',
  'car',
  'bus',
  'plane',
  'heart_pulse',
  'graduation_cap',
  'briefcase',
  'music',
  'zap',
  'wifi',
  'shirt',
  'dumbbell',
  'coffee',
  'gift',
  'book',
  'baby',
  'plus',
]

export const CategoryColors: CategoryColor[] = [
  'red',
  'orange',
  'yellow',
  'green',
  'teal',
  'blue',
  'purple',
  'pink',
]